local_tools/recon_all_clinical/local/requirements.txt
# Python requirements for tool:
# Ex: matplotlib==3.4.2
qmenta-sdk-lib==1.16.1.dev3412
pytest